test(push): move handlePush tests into KlaviyoPushOpenHandlerTest
